fixing my small change in the formula, sorry

+++ L2_GameServer_It/java/net/sf/l2j/gameserver/model/L2Attackable.java
@@ -1986,7 +1986,7 @@ public class L2Attackable extends L2NpcInstance
         if (Config.ALT_GAME_EXPONENT_XP == 0 && Config.ALT_GAME_EXPONENT_SP == 0)
         {
             if(diff > 5) // formula revised May 07
-            	xp = xp*Math.pow(5/6, diff-5);
+            	xp = xp*Math.pow((double)5/6, diff-5);
             
             if (xp <= 0)
             {
